improved optimalization and readability

+++ Sonar_STM32L476RG/Core/Inc/servo.h
@@ -4,7 +4,7 @@
 
 #define STEP 100
 
-bool is_max_angle_180(int i, volatile int pulse_width, volatile bool *state);
-void adjust_rotation_direction(int i, volatile int *pulse_width, volatile bool state);
+bool is_max_angle_180(uint8_t i, volatile int pulse_width, volatile bool *state);
+void adjust_rotation_direction(uint8_t i, volatile int *pulse_width, volatile bool state);
 void calc_current_position(volatile int pulse_width, volatile uint8_t *current_position);
